PlayViewMetrics now always recalculates values if updateWithRect:boardSize:() is invoked
The purpose of this change is to allow DebugPlayViewController
to trigger a recalculation if one of the parameters it manages
is changed. Because PlayView does not expose the internal
PlayViewMetrics object, the only way the controller can achieve
its goal is to trigger a pseudo frame change on PlayView by
invoking the public frameChanged() method. For this to work,
since the frame does not really change, PlayViewMetrics must
not check if the old/new values of the frame are the same.