no unnecessary change of "computer is thinking" state
it is not necessary to change the "computer is thinking" state in
GoGame if the computer plays a move and then immediately
goes on playing (e.g. in a computer vs. computer game); the state
change just causes unnecessary GUI updates (although the
updates occur too fast for UIKit to actually render them)