toolbar/statusline controllers now delay updates during long-running actions
On the simulator, some glitches in toolbar button visibility could be observed
(e.g. back button sometimes was not visible after the app launch was complete).
With this change the glitches are gone. Due to the delayed updates mechanism,
the final update now occurs in the main thread context. It is therefore reasonable
to assume that the root cause was that there were many updates happening in
a secondary thread's context during LoadGameCommand.