*very* basic solution for new iPad layout using UISplitViewController
This commit serves as the base for future work. the basic layout is working now,
but there's clearly room for much improvement (e.g. where to place the current
board position view, where to place the discard board position button, layout
bug when a new game is started, if orientation changes while game info view
is displayed the split view is not updated correctly etc.).

The main work has been done in PlayViewController. This is an overview:
- There is now a superview method for each of the subviews of the front-side view
- Obviously the setup and frame calculation methods for those subviews have been
  extensively revised
- Some unnecessary/obvious comments have been removed
- A few method bodies have been relocated within the source file