ItemScrollView can now change its orientation dynamically
This change originally was developed to implement a more dynamic view
layout on the iPad where a device orientation change would also cause an
orientation change in the board position list view. The iPad layout is now
more static and ItemScrollView no longer needs to change its orientation on
the fly. Still, the capability is too useful to throw away, so I am committing
this untested change anyway.