add wooden background
the image files have been made from this texture [1]. on 18 January 2013
I have sent an email to psdgraphics@yahoo.com, asking for "non-profit use"
permission. I have not received an answer, which according to the notes
on this page [2] implies that it is OK to use the image (the author says
that "I will not reply unless something is not OK").

[1] http://www.psdgraphics.com/backgrounds/light-wooden-background/
[2] http://www.psdgraphics.com/about/terms-of-use/