iOS 6 compatibility: rotate to portrait upside down on iPhone
In iOS 6 the supported orientations are determined by the intersection of the
orientations supported by the app's Info.plist - which before this commit already
contained portrait upside down - and the orientations supported by the root view
controller - which before this commit was a default UITabBarController that supports
all BUT portrait upside down in its implementation of supportedInterfaceOrientations.
The change in this commit therefore is to introduce a custom tab bar controller that
supports portrait upside down, and to add that controller to the applications's two
.xib files.

A side effect of touching the .xib files with this new version of Xcode is that, even
though the document format is not changed and remains at IB 3.1, Xcode
automatically performs substantial and opaque changes to the .xib files.