use special value -1 for "board position last viewed" on user defaults upgrade
when the app is upgraded to a new version and the user defaults are updated to the
corresponding new format, the "board position last viewed" property now gets a value
that will cause the app to display the last board position of the game when it is
launched for the first time after the upgrade. this is the behaviour from before the
upgrade when there was no "board position" feature, and thus fulfills the principle
of least surprise.

before the change in this commit, the app displays board position 0 after the upgrade,
i.e. the start of the game. for the unwary user who does not yet know about the new
"board position" feature, this is a big surprise - she thinks that she has just lost her
game because she does not see any stones! then, when she tries to make the first
move, she gets another surprise: the alert that warns that future moves will be discarded.