code cleanup: fix CommandProcessor object ownership policy
CommandProcessor was originally implemented with an object ownership policy in mind
that has its roots in C++: The client creates the command object, then submits it to the
processor which takes ownership of the command object and destroys it when it is done
with it. This policy, although workable, is non-standard for Objective-C, and potentially
confusing/misleading to other people reading the code. With this commit the non-standard
behaviour is fixed, and as a side-effect the corresponding warnings issued by Xcode's
analyze tool are eliminated.