fix crash: app delegate pauses game only if it is in an appropriate state
without this fix the app crashes when the device is locked (or goes into inactive state for any
reason) while a computer vs. computer game is already paused or has ended. this crash is
due to an earlier refactoring in PauseGameCommand which removed the initializer, and with
it some safeguards