open archived games in other applications (enhance #19)
sending archived games (.sgf files) via email is useful, but with a small
additional effort it is possible to open those game files in *any* app that
supports the SGF file type; tested with Panda-Tetsuki, but also DropBox.

as an additional bonus, the content of .sgf files can now also be copied
to the pasteboard.