EXPERIMENTAL: remove requirement that cellWidth must be an even number
this requirement leads to a large waste of screen space on the iPhone on a 19x19
board. removing the requirement makes maximum use of the available screen
space. testing in the simulator and on a device with a number of board sizes and
zoom scales did not seem to reveal any special anti-aliasing problems other than
those I already know about, so I decided to remove the cellWidth = even number
requirement as an experiment.

known anti-aliasing problems:
- grid line anti-aliasing occurs at certain zoom scales, but this occurs even if
  cellWidth is an even number
- stones at small sizes are fuzzy because of the use of scaled images; again
  there is  no connection to cellWidth