"stone distance from fingertip" now works correctly when zoomed
before this change, zooming the Play view did also scale up the calculated stone distance
from fingertip; this lead to ridiculous if not outright unmanageable distances.

the change fixes the algorithm for calculating the stone distance from fingertip so that it
no longer depends on the Play view's zoom scale; furthermore, the algorithm is now also
independent of the board size (which seemed such a clever idea when I started work on
this project).

as a result of the changes, the algorithm is now very simple and much more
understandable; as a consequence, the "stone distance from fingertip" user preference
value has changed its semantics and is now a percentage, but this is purely a technical
change, the user sees no change in the GUI.