no longer save app state when app is suspended (reverts 775f3e3)
the change made in 775f3e3 was FATALLY wrong: application state must not be saved during
application suspend because the application might be in an inconsistent state at this time.

for instance, scoring might be in progress (this was actually what made me notice the
problem), or GTP command that belongs to a larger, unfinished operation might be
processed at the moment.

this case shows that the concept for saving application state is fragile and needs to
be redesigned so that it is more resilient against implementation changes; if the
developer needs to constantly think HARD whether or not is safe to save state at
this moment and place in the code, something is guaranteed to go wrong again in
the future. I have added issue #124 to address the problem.

finally, the consequence of reverting 775f3e3 is that some application state
properties (board position, GoGameDocument properties, scoring mode state)
are not saved as often as possible. instead they are saved only after changes
that also trigger the .sgf file to be saved.