no longer write/read GoScore to/from NSCoding archive during backup/restore
restoring scoring mode from the NSCoding archive does not work if the archive is created
while scoring is in progress. this is exactly what happens, though: discarding a move while
in scoring mode 1) triggers a new round of scoring, then 2) makes a backup. if the backup
occurs while scoring is still in progress (ALWAYS happens when the GTP engine is finding
dead stones), the resulting NSCoding archive contains a GoScore object that is still in the
middle of calculating the score.

instead of implementing a clunky fix for the problem in GoScore, we don't try to restore
scoring mode at all. a real and better fix for the problem should hopefully emerge when
#124 is implemented in the next version.