retain current board position by saving the app state after each board position change
in ffeaf54 we removed the property that could be used to save/restore the current board position;
in 96b0eeb we had to remove the ability to save the app state on app suspension due to a fatal
flaw of the concept. as a result we are no longer able to restore the current board position after
an app launch. this is unacceptable because the user had this ability in the previous version of
the app, so the change in this commit is implemented as a temporary workaround.

the consequence is that changing the board position now takes a lot more time. for instance, on
an iPhone 3GS the backup takes about 1 second, on an iPad 2 it is about 0.3 seconds. although
this solution is rather unfortunate, I chose to implement it because it is only temporary and can be
reverted after #124 has been implemented in the next version.