guide user to enable logging if computer plays illegal move (cf. #90)
up until now the user was guided to send a bug report if the computer played an
illegal move. as it turned out, the bug report diagnostics file contained not much
valuable information on HOW the problem situation came into being, because
nobody had turned on logging at the time the problem occurred.

the new scheme is to guide the user so that she turns on logging instead of
sending a useless bug report. this should guarantee that logging is enabled
when the bug occurs next time. when that happens, the usual guidance to
send a bug report is provided.