make controllers aware of changes to player and profile objects
GameInfoViewController and PlayerProfileSettingsController can be visible at the same time,
and both controllers can initiate changes to player and profile objects that are tied to the
current GoGame. as a consequence both controllers must observe these player and profile
objects so that they can update themselves when a change occurs.

this commit completes the functional changes required to implement #122.