new GoGame property reasonForComputerIsThinking
we need to be able to distinguish between different reasons why the computer
is busy thinking (e.g. so that we can update the status view).

the old computerThinks property is now merely used as a convenience
property that simplifies querying the new property in those cases where a
simple boolean state is required.