remove shell script phase from unit test target
in Xcode versions prior to Xcode 5 it was necessary to have a shell
script phase for the unit test target that would execute the RunUnitTest
shell script; this shell script represented the unit test runner.

in Xcode 5, the RunUnitTest shell script is present, but it generates
an error saying „RunUnitTests is obsolete. To run unit tests for your
target, use the Test scheme action in the Xcode IDE and the test action
in xcodebuild.“

as a first step to make unit tests run again in Xcode 5, the shell
script phase is therefore removed from the unit test target.