remove Fuego and Boost dependency from unit test target
when running unit tests in Xcode 5 with the Fuego and Boost dependencies
active, there is a crash inside the Boost libraries at some point when
Boost attempts to destroy an exception object.

the reason for the problem is unknown. since I lack the time and the
inclination to deal with the issue, I decided to simply remove the
Fuego and Boost dependency from the unit test target.