execute unit tests when the „Test“ action is selected in Xcode (Cmd+U)
in Xcode 4 the unit test target had a shell script phase that made the
target runnable, i.e. it was possible to include the unit test target in
the „Run“ action of the main target. as a consequence, the unit tests
were conveniently executed each and every time that the main target was
run.

in Xcode 5 the shell script phase had to be removed (see previous
commit). as a consequence unit tests can no longer be executed when the
main target is run. the next best thing therefore is to execute them
when the „Test“ action is selected for the main target. although less
convenient than the original solution, this at least makes it
unnecessary to switch the target to run unit tests.