update submodule reference to re-created fuego-on-ios repository
it appears that the fuego-on-ios Git repository was set up incorrectly,
to the effect that it was impossible to synchronize with the upstream
SVN repository. because of this, the fuego-on-ios Git repository was
recreated from scratch.

this commit updates the submodule reference to the re-created
repository. the new reference points to a state inside fuego-on-ios
that includes upstream r1728 (the previous reference pointed to a state
that included r1709).