simplify board position table view selection handling on iPad (#198, #199)
this change is possible because we no longer have to support iOS 5 and
can now override tableView:shouldHighlightRowAtIndexPath:(). because of
that we no longer have to maintain each cell’s selectionStyle, and
because of *THAT* we don’t have to reload cells when the current board
position changes.

the behaviour described in issue #199 changes as a by-product of the
change in this commit: the „start of the game" board position table view
cell now has the wrong background color when the table view is scrolled
from bottom to top. the source of the problem, however, remains the
same and will be fixed in the next commit.