„start of the game" board position now has correct background color (fixes #199)
the main fix here is to *always* set the background color for all cells.
prior to this fix the background color was not set at all for the board
position 0 cell (i.e. the „start of the game“ cell). the cell reuse
mechanism of UITableView would therefore lead to an unexpected cell
background color.

because we now set the cell background color in
tableView:cellForRowAtIndexPath: instead of in
tableView:willDisplayCell:forRowAtIndexPath:, the latter method is no
longer needed.

because this also changes the timing of when the cell background color
is set, another smaller, unreported bug is fixed as a side-effect: the
cell detail text label sometimes and inexplicably had the wrong
background color. the cause for this was that setting the background
color in tableView:willDisplayCell:forRowAtIndexPath: simply happened
too late and that cell reuse again got the better of us.