revamp of EditTextController to use Auto Layout and the iOS 7 look
EditTextController now uses the plain white look of iOS 7 and no longer
sets up a gray table view background. Input elements are laid out
slightly differently due to this (especially the UITextView).

In addition to using Auto Layout, EditTextController now reacts to the
keyboard being shown and hidden when EditTextControllerStyleTextView is
in use. The UITextView constraints are adjusted dynamically in response
to the events so that the UITextView can grow/shrink.