add new custom view CoordinateLabelsView
the reason for adding a new custom view is that the view can participate
in the Auto Layout system by providing an intrinsic content size,
similar to PlayView, based on the values calculated in PlayViewMetrics.
this is much simpler than fiddling with the content size of the
containing scroll views.

structural changes that result from introducing the new custom view:
- PlayViewMetrics is now created by ApplicationDelegate instead of
  PlayView. The role of PlayViewMetrics has shifted towards a common
  model-like object, providing values required for drawing several
  views, instead of only a subsidiary object of a single view
  (PlayView).
- PlayView is no longer responsible for setting up the two coordinate
  label view instances and their container scroll views. This is now
  the task of PlayTabControllerPhone, which makes a lot of sense since
  the controller is responsible for the setup of the view hierarchy on
  the Play tab.