eliminate Auto Layout warning on Play tab (iPad only)
the warning was triggered because an intermediate view was created with
a frame of size zero. it is not entirely clear why this should be a
problem since initial frames should no longer matter - after all Auto
Layout is calculating the proper frames when layouting occurs.

nevertheless, the warning did occur, and changing the initial size did
fix the warning, so we don't fight the system but follow its guidelines.
the solution is to invoke

  [super loadView]

in the loadView override. the view that is thus created always has a
non-zero size. a different solution might have been to say

  self.view = [[[UIView alloc] initWithFrame:CGRectZero] autorelease];

but for the moment I decided to let UIKit do its magic, whatever it is,
instead of doing my own magic, which might prove to be wrong.