iPhone: display BoardPositionView instances using Auto Layout
until now the current BoardPositionView, and the ItemScrollView that
lists all BoardPositionViews, were displayed as custom bar button items.
this required all sorts of clumsy toolbar handling. the new approach is
to simply add those views as subviews to the toolbar and to show/hide
them when necessary.

to work well with Auto Layout, BoardPositionView now also provides
an intrinsic content size.