PlayViewMetrics refactoring, part 1: extract drawing functions
this is the first part of a series of refactorings whose ultimate goal
it is to make PlayViewMetrics into a proper model object.

we begin by removing all drawing functions from PlayViewMetrics and
instead place them into a new class aptly named PlayViewDrawingHelper.

at the same time, we move PlayViewMetrics to the "model" folder and
update the class docs to indicate PlayViewMetrics' new role.