PlayViewMetrics refactoring, part 3: fix "display coordinates" update mechanism
up until now the dependency of PlayViewMetrics on the "display
coordinates" user preference was an implementation detail not visible
to external clients. this caused clients to observe the wrong property
in PlayViewModel, with the result that clients would possibly get wrong
values from PlayViewMetrics.

this is now fixed: PlayViewMetrics publicly "admits" that it depends on
the "display coordinates" user preference by exposing both an updater
method and a property of its own. clients now use KVO on that property
instead of the one in PlayViewModel.