PlayViewMetrics refactoring, part 4: hardcode static drawing properties in PlayViewMetrics
originally, a number of static drawing properties were stored in the
user defaults, with the idea that these properties one day might become
real preferences changeable by the user. this never manifested, so now
I am finally hardcoding these properties.

additionally, the properties are now located in PlayViewMetrics instead
of in PlayViewModel. as a consequence of this second change, most of the
layer delegates no longer require a reference to PlayViewModel.