set contentsScale property on all CALayer objects to support Retina displays (#205)
the contentsScale property, which exists both on UIView and CALayer, is
an important property for properly supporting Retina displays. UIKit
automatically sets the property to its correct value on UIView objects,
but does not do so on CALayer objects - there we need to manually set
the property.

after this commit, coordinate labels are now displayed crisp and sharp
on Retina displays.