optimize drawing in CrossHairLinesLayerDelegate
the layer is now only drawn if something actually changed about the
cross-hair lines within the tile boundary (cross-hair lines have started
or stopped to intersect with the tile, or if they keep intersecting but
at a different position).

in addition, only that region of the layer is drawn that actually needs
a change (setNeedsDisplayInRect:).