optimize drawing in CrossHairStoneLayerDelegate
the layer is now only drawn if something actually changed about the
cross-hair stone within the tile boundary (cross-hair stone has started
or stopped to intersect with the tile, or if it keeps intersecting but
at a different position).

in addition, only that region of the layer is drawn that actually needs
a change (setNeedsDisplayInRect:).