BoardViewController no longer sync's the zoom scale of coordinate label scroll views
BoardViewController resets the zoom scale of the main scroll view to 1.0
after each zoom operation. this means that the coordinate label scroll
views always get sync'ed with a 1.0 zoom scale, so sync'ing is actually
not necessary at all (minimum and maximum zoom scales need not be
sync'ed because coordinate label scroll views are not interactive).

now that BoardViewController no longer needs to adjust the zoom scale
of coordinate label scroll views, it also no longer needs to be the
delegate of those scroll views. this also simplifies the implementation
of some UIScrollViewDelegate methods.