increase delay until long press gesture begins from 0.1 to 0.15 seconds
0.1 seconds is just a wee too short for my taste: on an iPhone 4S I
often accidentally place a stone when instead I want to scroll around
a zoomed board.

0.2 seconds is slightly too long, so the in-between 0.15 seconds seem
just right.

a slightly longer delay should also make it more clear to the casual
user that a stone cannot be placed by tapping. on the other hand, 0.15
seconds is not too long that the same casual user won't find out
intuitively that a long press is required.