update application icons for iOS 7 (downscale from appstore icon)
the previous version of the icons used a transparent border, in an
attempt to preserve the icon designer's rendering at the original size.
unfortunately, a transparent border is rendered on actual devices as a
black border, which looks very ugly.

while scaling an image down from a larger size is also not ideal (the
final icon might contain scaling artifacts), in this situation it is
clearly the lesser of two evils.