fix display of status view for long texts
the additional spacing between text and activity indicator in the status
view (added in commit 2b1d321) broke the view layout for long texts.

the change in this commit fixes the problem by dynamically
adding/removing the activity indicator to the status view. as long as
the activity indicator is not in the status view, long texts now benefit
from additional space. the tradeoff is that texts that are displayed
while the activity indcator is also in the status view must be relatively
short.