StonesLayerDelegate no longer crashes the app by creating array from nil object (fixes #242)
this commit includes two changes to fix the issue:

1) make sure that two member variables in StonesLayerDelegate can no
   longer get out of sync (self.currentCrossHairPoint and
   self.drawingRectForCrossHairPoint)
2) crude check for nil object before creating the array

the first change is a fix for the only potential flaw in the code logic
that I could find that might lead to the crash.

unfortunately, the identified flaw in the code logic is merely
theoretical, I was unable to deduct an actual problem from it that might
cause the crash. the second change therefore is a bit of defensive
programming that makes 100% sure that the crash can no longer occur.

obviously, with the root cause for the crash not identified, there might
be further crashes later on, but we'll just have to risk this. it is
much more important to quickly release a bugfix now because the crash
is a reality right now!