upgrade submodule MBProgressHUD to the current HEAD of master
the most recent versioned release 0.9 is about half a year old. in the
meantime there have been a few fixes regarding iOS 8 that should be
integrated. since no official release exists for those changes we have
no choice but to point the submodule at the unversioned HEAD of the
master branch. the most recent commits indicate that this is some sort
of pre-release of version 1.0.