update URL for submodule QuincyKit; upgrade submodule to the current HEAD of develop
the most recent versioned release 2.1.9 is almost 2.5 years old. in the
meantime there have been a number of important commits, most importantly
updates that enable 64-bit builds. since no official release exists for
those changes we have no choice but to point the submodule at the
unversioned HEAD of the develop branch. commits from January 2014
indicate that this is some sort of pre-release of version 3.0.