fix status view-related auto layout warning when rotating on iPad (fixes #262)
managed to fix this only by revamping the entire layout scheme in
NavigationBarController: we no longer embed a custom view into a
bar button item because custom views with varying sizes don't seem to
work properly in all cases. this has been a pain in the neck for a long
time now. instead we now simply create three navigation bars that
appear to the user as a single bar. see the new implementation for more
details.