add "launch image" mode to MainTabBarController
creating new launch images is always a bit of a painful process because
there are so many different variants. MainTabBarController now has a
special "launch image" mode that sets up a minimal fake UI that can be
used to take screenshots, which can then be further processed to make
the final launch images. the "launch image" mode can only be enabled by
changing a hardcoded value in the MainTabBarController initializer.