iPhone 6 Plus devices can now rotate to landscape (#253)
all layout decisions that were previously based on the user interface
idiom are now based on the new UIType enumeration and the new
LayoutManager class that were introduced in the previous commit.

to make the change in this commit simpler, iPhone 6 Plus devices for the
moment use the same UI as the iPad. obviously this is going to change
soon...