don't switch application root view controller on interface orientation change
a switch to an entirely different view controller hierarchy is way too
much work. so instead we keep the basic view controller hierarchy intact
and merely change the view / view controller hierarchy when the Go
board is visible.

for the moment we use PlayTabControllerPhone when in Portrait
orientation. this is likely to change in the future.

the responsibility for being the UINavigationController delegate had to
be moved from PlayTabController to the new class
PlayTabNavigationController, because otherwise PlayTabController will
interfere with MainNavigationController, who also needs to be a
UINavigationController delegate.