iPhone 6 layout for Portrait now uses the system provided navigation bar
for the iPhone 6 layout the plan is to show the status view separately,
i.e. not embedded into the navigation bar as the other layouts do. this
finally allows us to ditch the horrible kludge of manually managing a
navigation bar (actually THREE bars). there is still hope to get out of
the hell pit of semi-explicable bar handling that we live in since we
adopted iOS 7.

as a consequence, NavigationBarControllerPhone does not need to be a
UIViewController, a simple NSObject will do. as a further consequence,
NavigationBarControllerPhone cannot be a subclass of
NavigationBarController, so we split off the basic UIBarButtonItem
management stuff into a NavigationBarButtonModel.

this slowly starts to look sane, and we are getting closer to Apple's
model of separating view (UINavigationBar) from data (UINavigationItem).