save app state when scoring mode is enabled in response to the game ending
app state saving has never been done before for this particular action;
this was a bit of an oversight, but did not matter much in the "grand
scheme of things" because the user was able to freely switch between
scoring mode enabled and disabled at any time.

with the introduction of play resumption handling in the past few
commits, the fact that the app is in state "scoring mode enabled" has
now become very important and thus needs to be saved.

as a short reminder: being in the state "scoring mode enabled" has
become very important because the transition from scoring mode enabled
to disabled is now the trigger for resuming play.