add "resume play" game action
up until now the only possibility to resume play was by leaving scoring
mode after a game had ended - in that case the app would automatically
resume play.

this solution is not sufficient if the user is currently viewing an old
board position - in that case the app MUST NOT automatically resume play
because this would interfere with the "view old board position" mode of
the app.

the only sensible solution is to provide an action to the user that
allows him to manually resume play when he wants to.