in iOS 8, don't use UIActionSheet for "More Game Actions" because it's buggy
on iPhone devices, if the "More Game Actions" action sheet is shown
while in UIInterfaceOrientationPortraitUpsideDown, the action sheet
is displayed upside down. there seems to be no fix for the issue, but
since UIActionSheet is deprecated in iOS 8 anyway, and its replacement
UIAlertController does not exhibit the problem, the most pragmatic
solution is to use UIAlertController when it is available (check at
runtime).