replace all references to the "default GTP engine profile" (#281)
the expression "default GTP engine profile", or "default profile" for
short, should no longer be used because in the UI this profile is now
labelled "Human vs. human games".

consequently, this commit changes all places in the code where the
"default profile" is referenced, so that these places now refer either
to the  explicitly named "human vs. human games" profile, or to the
"fallback profile" if a short name is needed.