default mode for magnifying glass is now "always enabled" (#271)
when the app is upgraded, the board size currently in use is not known.
it might be 19x19, in which case the auto-enabled magnifying glass would
be visible immediately (good). but it might also be 9x9, in which case
the user would not notice the new feature (bad).

because existing users might have hated the old
"stone distance from fingertip" feature, it is important that they
notice the new magnifying glass feature. this change is made to
guarantee that the feature is noticed regardless of which board size is
currently in use by an upgrading user.