fix crash due to unexpected send of UIApplicationWillChangeStatusBarOrientationNotification
the crash fixed by this commit is the same as the one described in
ticket #287 in the comment titled "Crash 2". the crash can be reproduced
somewhat simpler than described in the ticket, like this:

1. launch app on any iPhone
2. rotate to UIInterfaceOrientationPortraitUpsideDown
3. tap "more game actions"
4. select "new game"
5. tap "cancel"

the app now crashes because in step 4 iOS unexpectedly sent
UIApplicationWillChangeStatusBarOrientationNotification, which causes
GameActionManager to deallocate GameActionsActionSheetController. this
is very bad because at that time the controller was modally presenting
NewGameController, so when NewGameController is dismissed in step 5 the
presenting controller is gone - oops!

the pragmatic solution is to restrict handling of
UIApplicationWillChangeStatusBarOrientationNotification to iPad only;
incidentally, that's also the device that the notification handler was
originally written for.