scoring is now - again - possible if a player has resigned
this commit reverts parts of the changes in commit 4b778b5; if a player
has resigned it should remain possible to perform scoring actions, with
the aim to check whether the resignation was in order.

new in this commit is that GoScore now adds the information to its
result whether one of the players has resigned. as a consequence, the
status view must be made slightly higher in iPhone 6+ landscape to
accomodate the longer status text.