change type of property "supportedInterfaceOrientations" in various classes
affected classes: WindowRootViewController, LayoutManager.

UIViewController declares the property "supportedInterfaceOrientations"
to have type UIInterfaceOrientationMask, but we used type NSUInteger.
although UIInterfaceOrientationMask is just a typedef for NSUInteger,
it's better to use the actual type that is actually in the declaration.

type mismatch detected thanks to a compiler warning emitted by the new
clang version embedded in Xcode 7.3.1