Cocoa Lumberjack is built (again) as static library
This partially reverts the changes in commit
2336b19fe3bd1a2f4830a2d0d8cf3589f999e7de. In that commit, Cocoa
Lumberjack was updated to a new version. In the same commit this project
was changed to use a framework build of Cocoa Lumberjack. As it turns
out, that framework build produces a dynamic library that can't be used
on iOS (on iOS all libraries must be statically linked).

Here we revert back to building and using the static library version of
Cocoa Lumberjack.