set build setting ENABLE_TESTABILITY = YES for the Debug configuration
doing this because Xcode 7 recommends it when opening the project.

apparently ENABLE_TESTABILITY allows unit tests to access private
protocols, and possibly perform other shenanigans that would not be
possible under normal circumstances. documentation of ENABLE_TESTABILITY
is scarce, so there is no precise information on the effects of this
build setting at the moment. since we enable the setting for Debug
builds only there is no chance of negative side effects in production.