set build setting ENABLE_BITCODE = NO for all configurations
the default for ENABLE_BITCODE is YES. using YES when not all
dependencies are built with bitcode generates a linker error.

we have to say NO because the Boost and Fuego 3rdparty software builds
do not (yet) produce binaries with bitcode. it is not clear whether
whether the binaries consumed from CocoaPods contain bitcode.