submit crash reports to Crashlytics according to user preferences (#293)
Crashlytics by default always submits crash reports. this is different
from the behaviour of the previous crash reporting solution, QuincyKit,
which honored a set of user preferences, where the user could decide
not to submit crash reports at all, or to submit them only after
confirmation.

to alter the default submission beahviour, Crashlytics requires special
handling by a delegate. this commit adds such a delegate. notably, the
delegate may display an alert that asks the user for confirmation
whether to submit the crash report. when we still used QuincyKit, this
alert was displayed by QuincyKit.