iPhone 6+: "Game Info" and "More Actions" buttons are now enabled after rotating to landscape (fixes #299)
up until now GameActionManger did not include the "Game Info" and the
"More Game Actions" actions in the dictionary of visible game actions
that it delivers to its  UI delegate. probably due to this inconsistency,
GameActionButtonBoxDataSource was implemented so that it hardcodes a lot
of stuff that deals with those two actions.

starting with this commit, GameActionManger now adds the two actions to
its dictionary of visible game actions. because of this the
implementation of GameActionButtonBoxDataSource can now be made much
more generic - no more need to hardcode special handling for "Game Info"
and "More Game Actions".

as a "side-effect" of the improved implementation, the bug goes away:
the initial enabled state of the two actions is now available during
initialization GameActionButtonBoxDataSource.