cleanup: remove some unnecessary table view row height calculations
in older versions of iOS it used to be necessary to manually calculate
table view row heights for cells that contain multi-line labels. these
days iOS is handling those calculations correctly all by itself, via
Auto Layout, so our code can be removed.

since no one is using the UiUtilies helper method
tableView:heightForCellOfType:withText:hasDisclosureIndicator:()
anymore, and it is unlikely that it will be needed again in the future,
that method can also be removed.