iPhone 6+: fix crash when rotating from portrait to landscape while board view is not visible (#298)
this crash did not exist in the 1.2.1 release of the app, so it must
have come into existence with the upgrade to the iOS 9.3 SDK.

the crash occurs in this scenario:

- hold iPhone 6+ device in portrait orientation
- launch the app directly into the main menu, i.e. without making the
  board view visible
- rotate device to landscape

the crash happens because a gesture controller is removing itself from
a BoardView instance that has already been deallocated. this is
surprising because the BoardView instance should not have been created
in the first place, and neither should the gesture controller instance.

as it turns out, PlayRootViewControllerPhone inadvertently creates the
BoardView object in its viewWillTransitionToSize... override because
it does not check the "isViewLoaded" property before accessing the
"view" property. adding the check fixes the problem.

from some of the comments/documentation in PlayRootViewControllerPhone,
it appears that when the VC implementation was written the
viewWillTransitionToSize... override was not invoked when the view was
not loaded, so something must have changed since then. a review of
PlayRootViewControllerPhone is in order (cf. #303).