remove UiElementMetrics methods that access the UIScreen bounds (#304)
these methods are no longer used, so it is safe to remove them.

the trigger for removing them is that knowing and using the screen size
for layout-related tasks is no longer safe, because on iPad devices the
user can now split the screen and have two apps run in parallel.